GTK: Move CairoCachedSurface from being owned by GtkThemeService to gfx::Image.
CairoCachedSurfaces are representations of a GdkPixbuf that live on the display
server instead of in process. GtkThemeService currently returns a
CairoCachedSurface for a (IDR#, Display). Instead, make gfx::Image keep
CairoCachedSurfaces as a representation and make CairoCachedSurface work on any
display, keeping the mapping to display server resources as an implementation
detail.

This has the benefit of:

- You don't need to go through GtkThemeService to access all resources,
  simplifying the code in some places and allowing sharing of static images
  across profiles.
- This will let us (in the future) remove several image conversions in the GTK
  port.

This is part 1. Part 2 will remove the custom CairoCachedSurface calls from
GtkThemeService.

BUG=106060
TEST=none
R=rsesek@chromium.org
TBR=ben@chromium.org

Review URL: http://codereview.chromium.org/8769017

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@112778 4ff67af0-8c30-449e-8e8b-ad334ec8d88c