Revert 112770 - Prevent a crash in the first run search engine selector.
It is possible to try and close the window now before the TemplateURLService is loaded. In this case we will not have a fallback search engine to select by default. So we will just reject attempts to close the window in this circumstance until it is loaded and a fallback choice has been set. It should be an edge case.

http://crbug.com/106078
TEST=unit test
Review URL: http://codereview.chromium.org/8771018

TBR=ben@chromium.org
Review URL: http://codereview.chromium.org/8790001

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@112804 4ff67af0-8c30-449e-8e8b-ad334ec8d88c