Consider redirects to data:// urls as cancelling a request and don't report conflicts
If an extension A tries to redirect a request to http://foo and an extension B tries to redirect a request to data://..., we consider the redirect of extension B as an attempt to cancel the request. As cancelling a request is considered to take precedence over redirects, this does not generate a conflict message any more.


BUG=106431
TEST=no


Review URL: http://codereview.chromium.org/8802017

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@113068 4ff67af0-8c30-449e-8e8b-ad334ec8d88c