Reverts a commit that caused ASAN failures, and 2 dependent commits.
The primary commit was 113249, the dependents were 113261, 113263.

This is a speculative revert, r113249 is by far the likeliest culprit in the blamelist of build http://build.chromium.org/p/chromium.memory/builders/ASAN%20Tests%20%282%29/builds/2325 which is where we started seeing the ASAN failures in question, will un-revert if it does not fix the problem.

TBR=jhawkins@chromium.org
BUG=none

Review URL: http://codereview.chromium.org/8832006

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@113387 4ff67af0-8c30-449e-8e8b-ad334ec8d88c