net: replace DnsRRResoler with TransportSecurityState in plumbing.
The DnsRRResoler has served its time and I don't have any further plans for it.

The TransportSecurityState will be used (in a future CL) to enforce public key
pinning at certificate verification time. Currently we only enforce it in
url_request_http_job.cc which means that we've already sent the HTTP request
(inc cookies) by the time we catch the problem. This was expeditious while we
fixed some other issues, but it's time to fix it.

BUG=none
TEST=compiles


Review URL: http://codereview.chromium.org/8692012

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@113410 4ff67af0-8c30-449e-8e8b-ad334ec8d88c