Fix detection of default G+ profile picture
Currently we assume that any profile picture URL that contains "AAAAAAAAAAI/AAAAAAAAAAA" represents a default profile picture.

It turns out that this is incorrect. There are valid profile pictures with that URL.

With the OAuth user info API we can simply detect default profile pictures by the existance of the "picture" field in the returned data. Accounts with a default profile picture don't have this field.

BUG=106444
TEST=Compiled and verified that unit tests passed.


Review URL: http://codereview.chromium.org/8801023

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@113487 4ff67af0-8c30-449e-8e8b-ad334ec8d88c