GTK: Fix crash on avatar bubble close
If the avatar bubble closed while opening a profile we would crash.

Here's the sequence of events:
   AvatarMenuBubbleGtk::OpenProfile
      ->SwitchToProfile();
          ...
            AvatarMenuBubbleGtk::OnDestroy()
                delete this;
      this->bubble_->Close();  <--- crash here because this is deleted

To prevent these kinds of errors I changed the OnDestory() code to use MessageLoop::DeleteSoon(). This ensures that there no AvatarMenuBubbleGtk functions are in the backtrace when the bubble is deleted.

BUG=106640
TEST=Reprorduced the crash. Applied my patch and verified that it no longer crashed.


Review URL: http://codereview.chromium.org/8889027

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@113717 4ff67af0-8c30-449e-8e8b-ad334ec8d88c