Fix bug where panels sometimes don't align on bottom.
This happens when a move animation starts while a resize animation, typically during creation is in progress. Now the animation starts from where the last one left.

Fixes a bug where from bounds and to bounds are reversed in the called StartBoundsAnimation in OnSizeChanged.

BUG=106573
TEST=Instructions in bug report.


Review URL: http://codereview.chromium.org/8879005

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@113746 4ff67af0-8c30-449e-8e8b-ad334ec8d88c