first cut at uber page
the subpages are iframes. On the C++ side, the uber WebUI keeps a collection of subpage WebUIs and proxies function calls to them. Calls into JS are directed to the correct iframe by setting a frame name.

The fact it's in an iframe is almost completely transparent to the options page (both from the C++ and JS side); you'll notice there are no changes to any options files.*

*exception: temporary command line changes

BUG=100885
TEST=chrome://uber should have an iframe with chrome://settings in it (fully functional)

Review URL: http://codereview.chromium.org/8889041

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@114152 4ff67af0-8c30-449e-8e8b-ad334ec8d88c