[Sync] Add HistoryService::DeleteURLsForTest
This is so that the sync performance tests don't need to wait for each
URL delete to hit the disk before moving on to the next one.

This is a speculative fix to stabilize the delete_typed_urls sync perf
test.  It also cuts down the runtime of the typed urls test from >20s to <10s or so.

BUG=107040
TEST=

Review URL: http://codereview.chromium.org/8890089

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@114458 4ff67af0-8c30-449e-8e8b-ad334ec8d88c