Sync integration test for authentication errors
http://codereview.chromium.org/8921025 added logic to take care of the
case where changing the gaia passphrase caused a sync client that was
already signed in to request a reauth, and if it was unavailable, go to
one of the OFFLINE states.

This patch adds a sync integration test to verify that a client does
indeed go to the OFFLINE state when the gaia password has been changed
elsewhere.

BUG=none
TEST=sync_integration_tests

Review URL: http://codereview.chromium.org/8931013

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@114483 4ff67af0-8c30-449e-8e8b-ad334ec8d88c