Wire up onEvent and put it to use when UdpSocket::Write blocks.
This is a checkpoint commit. It does the following:

- if Write() blocks, calls onEvent() when the underlying socket finishes its work.
- Demonstrates a working framework for sending events back to JS.

It does NOT do any of the following:

- Prove that the above statements are true.

In order to cause a UDP socket to block, we pretty much have to mock it, and I determined that I'd have to refactor my Socket class to let me put in a ClientSocketFactory for testing, which will then let me work DelayedSocketData in net/socket/socket_test_util.h into the mix and simulate Write blockage. After discussion with Antony, we agreed to break up this work into this CL and then a refactor/mock CL, which I'm working on now.

BUG=106802
TEST=none (see above).

Review URL: http://codereview.chromium.org/8896013

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@114544 4ff67af0-8c30-449e-8e8b-ad334ec8d88c