Fix AuthErrorTest for Mac
The sync integration test for auth errors was failing on the mac buildbots. It turns out that this was due to keychain prompts on mac, as we were trying to add a password form, due to a known issue with mac keychain.

This patch switches the datatype used by the auth error test to bookmarks, so that this test can be reenabled. The underlying issue with keychain will be fixed in another patch.

BUG=107611
TEST=SyncErrorTest.AuthErrorTest

Review URL: http://codereview.chromium.org/8954018

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@114669 4ff67af0-8c30-449e-8e8b-ad334ec8d88c