Badge the original profile's desktop shortcut and ensure it is unbadged if it is the last profile left.
We do this by fetching the default profile and avatar when the shortcut manager needs to make changes to the default shortcut.

We also correct a few places in the original ProfileShortcutManagerWin code to ensure it uses the last remaining profile's details rather than default details.

TEST=On a fresh install of Chrome, create a new profile. Ensure that both profiles have the correct badges.
BUG=106513,107435

Review URL: http://codereview.chromium.org/8931015

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@114814 4ff67af0-8c30-449e-8e8b-ad334ec8d88c