Revert 114929 - Standardize StringToInt{,64} interface.
These changes address issue #106655.  All variants of StringToInt have been
converted to use the StringPiece class.  One instance of conversion, in
chrome/browser/history/text_database.cc, required copying an underlying
string.  This is because the string type in question could use 8 or 16
bit characters depending on the OS type, and because StringPiece is not
implemented as a template, the code cannot specify whether to create a
StringPiece or StringPiece16.  This should be remedied in a future CL.

R=erikwright@chromium.org
BUG=106655
TEST=


Review URL: http://codereview.chromium.org/8921006

TBR=tedvessenes@gmail.com
Review URL: http://codereview.chromium.org/8984007

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@114930 4ff67af0-8c30-449e-8e8b-ad334ec8d88c