Revert 114992 - Switch to PostTaskAndReply for AccessTokenStore.
This removes the only usage of the CancelableRequest class in
content/, and is in preparation for moving the class back to
chrome/browser.

The ability to cancel requests to the AccessTokenStore was never used in production, so the interface now changes to a simpler non-cancelable one, and the browsertest GeolocationAccessTokenStoreTest.CancelRequest which tested the functionality is removed.

BUG=98716

Review URL: http://codereview.chromium.org/8996006

TBR=joi@chromium.org
Review URL: http://codereview.chromium.org/8992018

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@114997 4ff67af0-8c30-449e-8e8b-ad334ec8d88c