Fix extensions install procedure to use the user's temp directory for unzipping.
Extensions should use the users temp dir instead of temp dir inside the profile because if we have roamed profile writing there from the sandbox will fail.

There is a reason we don't copy directly to the target too:
this has been done like that in order to ensure that no extensions is copied only partially to its final location when the profile and temp are on separate volumes.

BUG=84045
TEST=manual

Review URL: http://codereview.chromium.org/8507012

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@114998 4ff67af0-8c30-449e-8e8b-ad334ec8d88c