Fix tab click/tooltip to utilize the correct tab
Clicking on the overlapping tab area will now activate
the highlighted tab (instead of either one).

The tooltip displayed in the overlapping area will be
the title of the tab which is highlighted. This is achieved
by moving the tool tip regions out of the overlapping
tabs and in to the tab strip. Overlapping NSViews &
tool tips are apparently buggy in Cocoa.

BUG=21911
TEST=Hover/click on overlapping areas in the tab strip
and see that the correct tab is activated / title is shown.


Review URL: http://codereview.chromium.org/8912013

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@115186 4ff67af0-8c30-449e-8e8b-ad334ec8d88c