[Sync] Add nigori node conflict resolution.
We merge the set of encryption keys, types, and non-encryption data in the
conflict resolver. If the server's keys are undecryptable we preserve them
until the user provides the passphrase, at which point we store back the
merged set of encryption keys. This way, once everything is resolved,
the default passphrase will be the one that was committed first, and all
clients will have all sets of encryption keys.

BUG=105615
TEST=sync_unit_tests. Manually attempt to set two different passphrases and
encrypted types on synced machines. Ensure everything works.


Review URL: http://codereview.chromium.org/8917031

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@115248 4ff67af0-8c30-449e-8e8b-ad334ec8d88c