Drop the "sync" mount option when mounting removable media.
The "sync" mount option was previously added to minimize the chance of
incomplete writes on removable media as there is no eject button in the
File Browser to safely unmount/remove media. However, that hurts the
write performance a lot. An eject button is now added to the File
Browser, so we should no longer use the sync mount option.

BUG=chromium-os:21749
TEST=Tested the following on Cr48:
1. Insert a SD card and make sure it is mounted by the File Browser.
2. Run "mount" from a root shell to verify that "sync" is not in the
   mount options for the mounted SD card.

Review URL: http://codereview.chromium.org/8937014

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@115408 4ff67af0-8c30-449e-8e8b-ad334ec8d88c