Fix MockKeychain implementation on Mac
When chrome tests involving passwords are run on Mac, the native Mac keychain
pops up dialogs that block automated tests from running. The MockKeychain class
was designed to be an in-memory keychain to substitute the native keychain.

The MockKeychain API requires that its storage data be mutable by classes
that consume the keychain. This resulted in the original implementation of
MockKeychain using contiguous blocks of dynamically allocated memory instead
of using a standard container. Unfortunately, one of the limitations of this
design was that passwords could be deleted only from the end of the storage,
and things like "add password A, add password B, delete password A" were
unsupported. This caused several sync integration tests that involved the
addition and deletion of passwords to fail.

This patch refactors the MockKeychain class to use a map<int, SecKeychainAttributeList>
and map<int, KeychainPasswordData> in place of the SecKeychainAttributeList* and
KeychainPasswordData* members that were previously in use. This allows for the
deletion of arbitrary password forms from the MockKeychain, and enables more
complex scenarios involving passwords.

It also fixes some bugs around uninitialized creation_time fields in password
forms, and re-enables all the sync password tests that were previously disabled
on Mac due to the bugs described above.

BUG=77993,89808,90460,105999
TEST=sync_integration_tests:*PasswordsSyncTest.*, sync_performance_tests:*Passwords*, unit_tests:PasswordStoreMacInternalsTest.*

Review URL: http://codereview.chromium.org/9021004

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@115852 4ff67af0-8c30-449e-8e8b-ad334ec8d88c