Fix and re-enable passwords sync perf test
The password sync perf test has been disabled on Mac for quite a while, because of the keychain prompts that used to pop up. The tests were made to use a MockKeychain implementation with http://codereview.chromium.org/9021004.

Yet, the tests were failing when several passwords were added because of a bug in the keychain, that would treat two usernames / domains / paths with the same prefix as equal, and therefore, treat them like duplicates in the keychain.

This patch fixes the bug by adding more stringent checks to the logic in MockKeychain that determines whether a password being added is in fact a duplicate or not. It also re-enables the perf test that was previously disabled.

BUG=105999
TEST=PasswordsSyncPerfTest.P0

Review URL: http://codereview.chromium.org/9052006

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@116002 4ff67af0-8c30-449e-8e8b-ad334ec8d88c