Make proxy exceptions list have a flexible height to eliminate scrollbars on dialog.
For the proxy settings dialog the exceptions list will size itself between 50 and 150 pixels according to the dialog height. On the settings page, since the panels are not given an explicit height the exceptions list will start at 50px and expand up to 150px to avoid causing a scrollbar on the settings panel.

BUG=107060
TEST=View the proxy configuration dialog and observe that the proxy exception list sizes itself to the available dialog space.

Review URL: http://codereview.chromium.org/9085014

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@116373 4ff67af0-8c30-449e-8e8b-ad334ec8d88c