Fix var hiding bug in sync's UIModelWorker
There's an extra declaration of error_info that hides the original
declaration.  We use the variable in the inner scope.  When the scope
ends, return the still-uninitialized variable from the outer scope.

This commit fixes that bug by removing the declaration in the inner
scope.

BUG=105994
TEST=


Review URL: http://codereview.chromium.org/9087020

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@116538 4ff67af0-8c30-449e-8e8b-ad334ec8d88c