Revert 116512 - SwitchTabAndHideAutofillPopup failing linux chromeos bot.
First pass for new GTK Autofill popup.

This first pass just displays the new Autofill popup and then hides it at the 
correct times. The box is the correct size, but none of the text is drawn at the
moment. This is only enabled when the --external-autofill-popup flag is enabled.

BUG=51644


Review URL: http://codereview.chromium.org/8890035

TBR=csharp@chromium.org
Review URL: http://codereview.chromium.org/9109031

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@116575 4ff67af0-8c30-449e-8e8b-ad334ec8d88c