Track session storage mismatches when committing Instant previews.
Say the user is on a tab with a SessionStorageNamespace (SSN) id of S. The user
causes an Instant preview to be committed onto the tab. Currently, Instant
always creates the preview with a new, different SSN (say R). When the preview
is committed, S is lost (being replaced by R), which is a problem when the user
tries to go "Back" in the tab. Session storage is not guaranteed to persist, so
this is not a critical bug, but it's worth fixing.

What if we created the preview with the same SSN (viz., S)? This might work,
but if the preview is subsequently committed on top of a different tab, we have
the same problem as before.

This CL measures how often this (commit onto a different tab) happens, so that
we can judge whether the naive solution above is enough, or whether a more
complex solution is needed (such as keeping track of whether the preview has
accessed session storage at all, and if not, rebind its SSN dynamically when
committing).

This CL does not introduce any change in behaviour. Specifically, it doesn't
actually set the preview's SSN to the originating tab's SSN. It just tracks it
in a different variable to report histograms. Hence, no tests.

BUG=101946
TEST=none


Review URL: http://codereview.chromium.org/9108023

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@116741 4ff67af0-8c30-449e-8e8b-ad334ec8d88c