Hook up the SequencedWorkerPool to the browser thread.
This does some refactoring of the static data in the browser thread so we only have one global object instead of a bunch fo separate arrays.

It also hooks up the visited link master's I/O to use this new system as a proof of concept.
Review URL: http://codereview.chromium.org/9065009

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@116816 4ff67af0-8c30-449e-8e8b-ad334ec8d88c