Revert 116903 - On Windows 8 if the user touches the omnibox on a touch screen the On-screen keyboard(OSK) will now display immediately.
Creates an ITextInputPanel object and binds it to the omnibox and then uses it to display the keyboard when we receive a WM_POINTER event.

Only works on Win 8 since we need to use WM_POINTER to avoid issues with focus handling.

BUG=105637
TEST=Verify 105637


Review URL: http://codereview.chromium.org/8662018

TBR=abodenha@chromium.org
Review URL: http://codereview.chromium.org/9155002

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@116909 4ff67af0-8c30-449e-8e8b-ad334ec8d88c