Revert 117078 - Move creation and ownership of ResourceDispatcherHost and PluginService to content. This gives a few benefits: -avoid having each embedder know when to create/destruct these objects, as well as contained objects (i.e. those related to downloads) -avoid having to tell embedders about specifics of BrowserThread startup/shutdown -move ResourceDispatcherHost's getter to content where it belongs
Some code (extensions+promos) used the fact that RDH is NULL in unittests as a signal to not use the utility process. I've switches those unittests to set a flag on the objects instead.

I've taken out the DnsParallelism field trial (not used anymore, confirmed with jar) as it was the only thing that caused MetricsService to depend on IOThread initialization, which also depended on MetricsService (through FieldTrials). This two-sided dependency always annoyed me and made the code hard to restructure.

BUG=98716
Review URL: http://codereview.chromium.org/9150016

TBR=jam@chromium.org

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@117096 4ff67af0-8c30-449e-8e8b-ad334ec8d88c