Detect sync server communication errors
This is part 2 of 3 in a series of patches to remove the unnecessary
SYNC_CYCLE_CONTINUATION sync cycle that follows every commit.

This patch adds infrastructure to report errors involving communication
with the sync server back to the SyncSession and SyncScheduler.

It modifies SyncerProtoUtil::PostClientToServerMessage to have it return
an error code describing why it failed.  This function's callers (all of
which are SyncerCommands) use the infrastructure added in the first
patch of this series to return the error value to SyncShare().  From
there, the return values are placed into the session's StatusController
object.  The method SyncSession::Succeeded() method provides an easy way
to determine if the anything went wrong with the current sync session,
and its response is based on the stored return values.

This change has no effect on the client's behaviour.  Although it gives
the client access to lots more information, the client does not use it
for any decision making.

BUG=94670
TEST=


Review URL: http://codereview.chromium.org/9158004

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@117285 4ff67af0-8c30-449e-8e8b-ad334ec8d88c