Convert DirectoryBackingStore to new sqlite API
This includes a whole bunch of simple changes to update to new method
names and signatures, like s/bind_int/BindInt/.  It also includes some
less obvious changes that deserve mentioning:

- It removes support for allowing different threads to access the
DirectoryBackingStore.  This functionality is no longer used, and
removing it allows us to use a single database connection for the entire
lifetime of the DirectoryBackingStore.

- The new sqlite API does not support the "BEGIN EXCLUSIVE" transactions
that were used by the old DirectoryBackingStore.  However, with the
above changes and the new API's set_exclusive_locking() call we can
ensure that no other program/thread can access the database throughout
the lifetime of the DirectoryBackingStore object, not just while a
transaction is in progress.  This is a stronger guarantee and more
accurately reflects the assumptions made by this class.

- Other subtle changes to the internal settings triggered by the change
in API include:
  - "PRAGMA secure_delete = ON" is set by sql::Connection.
  - "PRAGMA synchronous = 2" is no longer explicitly set, though it is
    the default according to the sqlite docs.
  - "PRAGMA fullfsync = 1" is not set.
  - page size is now set to 4K (only applies to new databases)

- We no longer run "PRAGMA integrity_check(1)" on database load.  This
will be done as an SqliteIntegrityTest introduced in a separate CL.

- There was a harmless bug in the call to set the value of
notification_state in CreateTables.  It has been fixed.

BUG=77634,109668
TEST=


Review URL: http://codereview.chromium.org/8745022

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@117328 4ff67af0-8c30-449e-8e8b-ad334ec8d88c