Allow HtmlDialogUIDelegate to specify dialog ModalType.
The aura stacking window manager uses the ModalType of widgets to determine if they are to be placed on top of all other layers such as the login screen. The login screen makes use of several HtmlDialogs which must use MODAL_TYPE_SYSTEM in order to be stacked correctly.

TBR=estade,sail,abodenha

BUG=109673
TEST=Verify that dialogs such as proxy settings can be displayed over the login screen.


Review URL: http://codereview.chromium.org/9147021

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@117386 4ff67af0-8c30-449e-8e8b-ad334ec8d88c