Substitute ResourceHandlers with dummy ResourceHandler while request is transferred
This CL implements a new ResourceHandler that only allows cancelling its
request. The idea is that in case a URLRequest is being transferred to a new
renderer, we need to get rid of its old ResourceHandlers as these may have
dependencies to the old renderer which is about to die. We replace the
request's old ResourceHandler with a new TransferredResourceHandler until
the request is resumed by the new renderer.

BUG=107692
TEST=no


Review URL: http://codereview.chromium.org/9113028

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@117658 4ff67af0-8c30-449e-8e8b-ad334ec8d88c