A browser test created an echo server, waited for it to start listening, ran its tests, sent it a message asking it to quit, and then sometimes IMMEDIATELY KILLED IT without giving it a chance to process the quit message. This triggered an assertion failure in the server destructor that was largely harmless but good for the purpose of pointing out this programming error.
I fixed the problem by adding a WaitableEvent in the echo server that gets
signaled when cleanup is complete. Then the browser test is able to wait
on that event (with waiting on the UI thread and signaling on the IO thread)
before letting the server go out of scope.

BUG=109337
TEST=already covered.


Review URL: http://codereview.chromium.org/9213005

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@117692 4ff67af0-8c30-449e-8e8b-ad334ec8d88c