Cloud Print Connector Policy, Service-side enforcement.
Have the service process periodically launch the browser (if not connected to
the browser already) with a command line switch that instructs the browser to
check the policy, shut the service down if it's set, then quit. If the service
is already connected to the browser, the launch is not done since the browser
is already watching for policy changes and would immediately inform the
service.

BUG=98049
TEST=Launch the browser, sign into the cloud print connector, quit the browser,
set the policy, wait a long time (8 hours) or relaunch the service process -
the service process should end up disabled and shut itself down.


Review URL: http://codereview.chromium.org/8905023

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@117732 4ff67af0-8c30-449e-8e8b-ad334ec8d88c