Create Function to Fix NTP Ordinal Collisions
This function will handle repeated values in the NTP by adjusting the elements that have collided and arranging them in order of increasing id. This end result will be the same regardless of the order the repeated elements are in, so different instances of chrome will have the same result (after getting colliding values after a sync).

BUG=61447
TEST=unit_tests:ExtensionSortingFixNTPCollisions*


Review URL: http://codereview.chromium.org/9230010

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@117909 4ff67af0-8c30-449e-8e8b-ad334ec8d88c