Added explanatory dialog for enterprise users during OOBE.
This is only shown when auto-enrollment is performed and doesn't succeed
automatically at first. It explains the user what he has to do, and gives a
(buried) option to opt out of enterprise enrollment.

This feature is behind a flag, and is still disabled by default.

BUG=chromium-os:23063
TEST=Login with an invalid (e.g. @gmail.com) account at the first sign-in screen during OOBE. Auto-enrollment should fail and show an explanatory dialog.


Review URL: http://codereview.chromium.org/9022049

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@118087 4ff67af0-8c30-449e-8e8b-ad334ec8d88c