WebUI TaskManager: Delay the task list update.
Sometimes, tabs/tasks are frequently updated and a lot of update messages come in a short time. In such case, the list 
trys to update UI again and again, the UI stops and does not accept any input for a while. 

To avoid it, this cl delays the update of the list and combine updates if they come within a second.

BUG=110424
TEST=manual


Review URL: http://codereview.chromium.org/9221009

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@118132 4ff67af0-8c30-449e-8e8b-ad334ec8d88c