aura/chromeos: Avoid suspending while video is playing.
This adds an ash::VideoDetector class that watches for layer
updates and attempts to detect the playback of video.  A
Chrome OS-specific observer updates a _CHROME_VIDEO_TIME
property on the X root window, which is used by the power
manager to defer screen dimming or other power management
features.  This matches the implementation currently in use
on Chrome OS, with Chrome taking the X window manager's role
for detecting video and setting the property.

BUG=110114
TEST=added; also manually checked that the property is updated


Review URL: http://codereview.chromium.org/9249004

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@118171 4ff67af0-8c30-449e-8e8b-ad334ec8d88c