Close button is not effective when a panel slides into place.
Details: When you have multiple panels of the same size and you close the right most panel, all remaining panels slide to the right. The close button of the right most panel will now be right under the mouse. In this scenario clicking on the close button does not have any effect until you move the mouse. This creates a bad user experience when the user tries to close multiple similar sized panels (typical scenario) quickly.

Root cause: In enter-notify-event handle for the panel window, there is a gtk_widget_show() call to panel wrench button. This seems to be getting in the way of the enter-notify-event being propagated to the close button which in turn is leaving the close button un-clickable.

Fix: When a panel finishes a move or size animation, if the mouse is over the close button, send an enter-notify-event to make the close button clickable.

BUG=107039
TEST=See bug report for repro.


Review URL: http://codereview.chromium.org/9179001

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@118219 4ff67af0-8c30-449e-8e8b-ad334ec8d88c