Merge Cloud Print signin dialog into regular Cloud Print dialog.
This is stage 1 of a 2 stage refactoring.  This combines the two dialog implementations.  The flow here is still not ideal but I'll address it in stage 2.

Stage 2 is more of a re-implementation than refactoring.  I will remove the cloud print dialog entirely, instead merging it into the print preview implementation.

BUG=109084,109283
TEST=Verify bugs


Review URL: http://codereview.chromium.org/9139082

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@118348 4ff67af0-8c30-449e-8e8b-ad334ec8d88c