Cleanup: ResourceDispatcherHost::BeginDownload takes a scoped_ptr<URLRequest> argument.
This used to take an URLRequest*, with a comment about ownership. Changing to a scoped_ptr argument makes the ownership explicit and enforceable by the compiler.

Also, changed the name of the child_id argument to be process_unique_id to be consistent with the rest of the file.


Review URL: http://codereview.chromium.org/9150028

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@118351 4ff67af0-8c30-449e-8e8b-ad334ec8d88c