Hack: remove dangling chrome://theme/IDR_CLOSE_BAR* references
In issue 9187061 I stupidly overlooked some references to these moved assets
(close box images). This change will cause them to appear again by replacing
the chrome:// references with paths to the PNG images. These paths are turned
into data references during the build process, which will add a few kilobytes
to the executable size (the images are 125-227 bytes each).

What should be done instead is to update the ThemeSource to find the assets
in the new location. This is a temporary fix until that happens.


BUG=110824,110828
TEST=none


Review URL: http://codereview.chromium.org/9192037

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@118520 4ff67af0-8c30-449e-8e8b-ad334ec8d88c