Revert 118649 - Move SigninManager initialize to factory.
Reverted because it caused a perf regression, as expected.

This is the change I think that introduces the startup perf regression - I'm landing it as a separate patch (split out from the previously r117130) to verify that this indeed caused the regression. If so, I'll revert it.

BUG=108017
TEST=watch perf bots
TBR=tim@chromium.org

Review URL: https://chromiumcodereview.appspot.com/9273003

TBR=atwilson@chromium.org
Review URL: https://chromiumcodereview.appspot.com/9270037

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@118653 4ff67af0-8c30-449e-8e8b-ad334ec8d88c