Fix flashing window when deleting profile
Deleting a profile with the sync promo open would cause the delete profile's browser window to temporarily re-open.

The problem was that the sync promo was attempting to navigate to a page when the promo was closed. Fix was to disable the navigation if the browser window was being closed.

BUG=108801
TEST=


Review URL: http://codereview.chromium.org/9193004

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@118710 4ff67af0-8c30-449e-8e8b-ad334ec8d88c