Initial draft of chrome://omnibox/ page.
Takes input text and display information about autocomplete suggestions.
There is an abbreviated mode and a detailed mode.
There is a mode to show all intermediate autocomplete updates and a mode
to show only the final suggestions.

This page is listed in chrome://about

This page is not internationalized.  Because it's only for debug purposes,
I don't think it needs to be.

There are a lot of ways to improve this.  We can display more information.
We can make the URL save-as-able.  (Right now if you do save-as, you see
the empty page, not page with any content.)  There are lots of things that
can be done, but I think this is good enough as is that it's worth
checking in.

BUG=79883
TEST=


Review URL: http://codereview.chromium.org/8888030

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@118772 4ff67af0-8c30-449e-8e8b-ad334ec8d88c