Rename visual style "password" to "obscured"
This change renames some identifiers to distinguish the visual style of being drawn **** **** from the logical property of being a password. It is split off from issue 8748001 and issue 8747001. See http://codereview.chromium.org/8748001/#msg12 and http://codereview.chromium.org/8748001/#msg14 for discussion of the name.

TEXT_INPUT_TYPE_PASSWORD is unchanged; it should perhaps be renamed, but not to OBSCURED since it affects input, not display.


BUG=none
TEST=It compiles. Ship it.


Review URL: http://codereview.chromium.org/9117012

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@118796 4ff67af0-8c30-449e-8e8b-ad334ec8d88c