[Sync] Add position change detection in conflict resolver.
We compute the prev id based on the server position and compare it to the
prev id from the client view. We use this in determining whether or not to
step into the "everything matches" case, or continue to the other conflict
resolver scenarios.

BUG=110903
TEST=sync_integration_tests --gtest_filter="*RacyPosition*"

Review URL: https://chromiumcodereview.appspot.com/9138043

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@118914 4ff67af0-8c30-449e-8e8b-ad334ec8d88c