Failover to omnibox search if the Instant loader isn't ready.
If the loader isn't ready or has a failed HTTP status code, make use of that to
fail over to a regular omnibox search in the HIDDEN and SUGGEST field trials.
The visible INSTANT trial already does so. Unfortunately for the SILENT trial,
since the first Update happens only at commit time, we can't use this logic.

BUG=110962
TEST=none

Review URL: https://chromiumcodereview.appspot.com/9283032

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@118976 4ff67af0-8c30-449e-8e8b-ad334ec8d88c