[Sync] Improve calculation of number of default entries to be more automated.
We now calculate the number of default entries by checking the number of total
entries on client 0 after sync setup and subtracting aware the datatype-specific
top level folders as well as a hardcoded value corresponding to other permanent
items that exist.

In order to support this, we now ensure that all clients are in quiescence before
we consider sync to have finished setting up.

BUG=97798
TEST=sync_integration_tests


Review URL: http://codereview.chromium.org/9212057

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@119004 4ff67af0-8c30-449e-8e8b-ad334ec8d88c