Revert 119018 - Draw new tab button with tab background
Currently the new tab button is drawn with a semi-transparent image blended with the window background color.

This CL changes the drawing code so that we now use the inactive tab background image as the button's background. I've removed the fill from the newtab button images so that they just contain the border and shadow.

The main benefit of this approach is that the button now matches the look of inactive tabs.

Note, I'm only implementing this on views for now. Once this is implemented on all platforms we can remove the old button images.

Screenshots: http://imgur.com/a/iez8r

BUG=100775
TEST=
Review URL: https://chromiumcodereview.appspot.com/9252026

TBR=sail@chromium.org
Review URL: https://chromiumcodereview.appspot.com/9288015

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@119020 4ff67af0-8c30-449e-8e8b-ad334ec8d88c