Separate the generated date from xkeyboard.cc for ease of udpate.
Currently we need to be careful not to delete the start and the end marker lines so that gen_keyboard_overlay_data.py work properly. This change move the generated part of xkeyboard.cc to a different file in order to make it easy to update the data and xkeyboard.cc.

BUG=None
TEST=unit_tests --gtest_filter="XKeyboardTest.*" passes

Review URL: http://codereview.chromium.org/9289044

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@119293 4ff67af0-8c30-449e-8e8b-ad334ec8d88c