Enable NTP Application Position and Page Name Syncing
This enables the icon on the NTP to have their positions synced so that users will always have there icons appear in the same place. The NTP page names are also synced, since the required number of pages will correctly appear for each instance though the app syncing.

BUG=61447, 100737, 111277
TEST=sync_integration_tests:TwoClientAppsSyncTest.*
Users should be able to move icons on their NTP or change the page names and have the results propagate to there other signed in instances.
If a user deletes an app from one client while another client NTP is open, the second client's NTP should update to  look the same as the first's.


Review URL: http://codereview.chromium.org/9254039

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@119298 4ff67af0-8c30-449e-8e8b-ad334ec8d88c