Profiles: Move FindBarState to PKSF.
This patch is absurd. Previously, we had a bunch of boilerplate code
in Profile to manage a class whose only member was a single
string. Now, we have a different set of boilerplate code and build
static factories to manage the lifetime of said object.

And yet, FindBarState is exactly the sort of thing PKSF was built for:
it's an object that hangs off the Profile that some chromium variants
probably don't want.

BUG=77155


Review URL: http://codereview.chromium.org/9271080

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@119327 4ff67af0-8c30-449e-8e8b-ad334ec8d88c