Fix logic that determines if a user has typed a custom passphrase during setup
The chrome histograms are revealing an unexpectedly high number of sync
users who use a custom passphrase. It turns out that this is because we
are also counting sync configurations where the value of
set_gaia_passphrase is true.

This patch fixes the logic that detects when a custom passphrase has
been set.

BUG=111729
TEST=set up sync without going to the advanced dialog and check the histograms generated

Review URL: https://chromiumcodereview.appspot.com/9296024

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@119532 4ff67af0-8c30-449e-8e8b-ad334ec8d88c