Fix WiFi & VPN connection dialogs when controls a disabled through policy.
Previously, the UI would make decision based on the enabled() state of
controls, indicating that the control is not relevant for the currently
chosen configuration. This doesn't work for policy-configured networks,
since the control might just be locked down, in which case it is
relevant, but disabled.

BUG=chromium-os:23124
TEST=Configure ONC networks through policy, observe that the connection dialogs show the configured values even though disabled.

Review URL: https://chromiumcodereview.appspot.com/9296009

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@119664 4ff67af0-8c30-449e-8e8b-ad334ec8d88c