[uber page] Fix the missing cancel button on the settings search field.
On Mac, the search field's cancel button appeared to be missing. This was caused by applying
a transformX() to header elements to allow them to side-scroll properly. This CL changes to
use margin-left instead. The behavior is the same, but now the cancel button appears properly.
I will be filing a WebKit but since this is probably a core issue with the search control.

BUG=112046
TEST=Verify that search field displays a cancel button when not empty.
Review URL: https://chromiumcodereview.appspot.com/9318009

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@119999 4ff67af0-8c30-449e-8e8b-ad334ec8d88c