uber settings - checkbox/radio clickability
change labels to -webkit-inline-box. This doesn't change the children but the box itself will behave like an inline element rather than block (won't expand horizontally). Also remove box flex on the span as it's not necessary. Also add some padding on the end of the labels so that there is extra clickable area when the text is short (this doesn't change anything if the span is long enough to wrap).

In native dialogs:
On Linux/GTK, the clickable area extends all the way to the right (our behavior before this patch).
On Mac, the clickable area is only the text and the checkbox
On Windows, the clickable area extends beyond the text but not all the way to the far right.

I'm splitting the difference by matching Windows native behavior, which seems both logical and the middle road.

BUG=112091
TEST=1. Checkbox labels should be clickable on either the text or the checkbox, but not more than 100px from the right edge of the checkbox. 2. Long labels still wrap with indent (see bug 109334)

Review URL: https://chromiumcodereview.appspot.com/9307015

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@120033 4ff67af0-8c30-449e-8e8b-ad334ec8d88c