Fix Idle NTPs to Update
Fix the extension code so that everytime a icon is moved it sends a notification. It was only sending notifications when the icon was moved and it had to get its ordinal value (which only occured if it was moved to a page that had at least 1 other icon).

BUG=110269
TEST=When two NTP are open, users should be able to move one icon to a new page and have the change reflected in both views


Review URL: http://codereview.chromium.org/9307066

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@120332 4ff67af0-8c30-449e-8e8b-ad334ec8d88c