chromeos: bluetooth: tie Proxy lifetime to object, not observer
Change the lifetime of Object Proxies to the remote object, not any
Observer; one of the problems with the previous method was adding
a second observer would treat every new attempt to add an observer
as a new proxy so reconnect signals, etc.

The other main problem is that you couldn't call methods unless there
was an observer that had created the object proxy.

BUG=chromium-os:22086
TEST=verified methods still work

Change-Id: Ic58b88bbad84c7728abd16ddc48d11174f59f59f


Review URL: http://codereview.chromium.org/9314013

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@120456 4ff67af0-8c30-449e-8e8b-ad334ec8d88c