uber extensions
change trash to be a button so that clicking it doesn't focus it (as I recently discovered).

the problem described in the bug stems from the fact that clicking the div focuses it, and a focused trash icon is visible and open (to support keyboard usage).

BUG=112496
TEST=see bug

Review URL: https://chromiumcodereview.appspot.com/9323052

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@120674 4ff67af0-8c30-449e-8e8b-ad334ec8d88c