We make the server return error during the sync setup(configuration) and we make sure the SetupSync on client succeeds when the server returns success.
The server returns a transient error on the first 2 requests of every 3 requests. So when downloading nigori we get 2 errors first and then a success. Similarly when downloading other types we get 2 errors before succeeding.

Note: In test cases there is no wizard because of which we dont throw an unrecoverable error and instead wait. So setup actually succeeds when server returns success.


BUG=100076
TEST=try bots


Review URL: http://codereview.chromium.org/9290060

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@120878 4ff67af0-8c30-449e-8e8b-ad334ec8d88c