Aura: Fix activated windows sometimes not switching to foreground
This was caused by a mismatch in window and layer stacking.  We use a NULL layer delegate as a signal that a window is hidden or closed and is animating out.  The bottom-of-window status bubble frequently shows and hides, and hence is commonly in this state.  If you tried to activate a background window while the status bubble was animating out the window would activate and start processing events, but its layer would not be brought to front, so it would visually appear behind other windows.  The problem was probably introduced in either http://crrev.com/119453 (which introduced stacking changes for NULL layer delegate windows) or http://crrev.com/119753 (which fixed a similar issue with transient child windows).  This patch ensures that window stacking order always matches layer stacking order and adds unit tests for this situation.

BUG=112562
TEST=aura_unittests WindowTest.Stack*

Review URL: https://chromiumcodereview.appspot.com/9371028

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@121351 4ff67af0-8c30-449e-8e8b-ad334ec8d88c