Fixes login screen menus.
Pass in an instance of a window to use for testing whether the inner dispatcher can receive events.
The |associated_window| should be such that it is behind a window which stops event propagation whenever the inner dispatcher should not receive events.

This fixes menus on the login screen because menus are above the lock screen in z order.

Bug=113247
Testing=Manual

Review URL: http://codereview.chromium.org/9374028

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@121570 4ff67af0-8c30-449e-8e8b-ad334ec8d88c