Add client for background testing of HTTP pipelining.
Note this isn't hooked up, so it won't affect users yet. Once the servers are up, it'll be
hooked up to a field test, much like the UDP echo test. Also missing is a force-pipelining option for requests. I'll add that in a separate CL.

BUG=110794
TEST=unit_tests


Review URL: http://codereview.chromium.org/9302024

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@121572 4ff67af0-8c30-449e-8e8b-ad334ec8d88c