Increase the rate at which we check
This is because now that we have several components available, whenever we successfuly download
one we would sleep for 4 hours, no matter if other components could be installed.

With this change we only sleep for 1 hour, once all components are up-to-date or have no updates we
check only after MinimumReCheckWait()

So as long there are no components to install the only thing that this change does is to
post a task to the UI four times as fast as before, e.g. every hour.

BUG=113749
TEST=see bug.

Review URL: http://codereview.chromium.org/9380031

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@121600 4ff67af0-8c30-449e-8e8b-ad334ec8d88c