[Sync] Ensure we don't access the ProfileSyncFactory after sync shutdown.
Previously, the ProfileSync[Components]Factory was owned by the profile itself,
and we could rely on it persisting while the profile still existed. Now that
the PSS is a ProfileKeyedService, the PSS itself owns the factory. As a result,
NewNonFrontendDatatypeControllers can no longer rely on the factory being
valid when beginning association.

We now remove the GetWeakPtrTosyncableService method of the DTC in favor of
having that functionality be generic and done within the SharedChangeProcessor.
This allows us to both simplify the DTC and ensure we don't access the factory
if we've been disconnected (which we always are at shutdown time).

BUG=103407
TEST=unit tests


Review URL: http://codereview.chromium.org/9369043

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@121608 4ff67af0-8c30-449e-8e8b-ad334ec8d88c