[uber extensions] flicker less when enabling/disabling extension
the real fix will be to make the extensions page not clobber several times in a row when an extension is disabled/enabled, but that is more difficult. (See comment in extension_settings_handler.cc::Observe). For now this makes it look a bit better.

BUG=112493
TEST=see bug


Review URL: http://codereview.chromium.org/9381026

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@121633 4ff67af0-8c30-449e-8e8b-ad334ec8d88c