[uber] fix lazy load in caching case
the iframe contents can be cached even when src= is not set. The easiest way to tickle this bug is to set chrome to restore session on startup, with chrome://chrome as one of the pages. Workaround: dynamically add the iframe rather than just setting the src attribute (suggested by dbeam)

BUG=none
TEST=none


Review URL: http://codereview.chromium.org/9379031

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@121641 4ff67af0-8c30-449e-8e8b-ad334ec8d88c