Remove ssl_host_state.h dependency from chrome.
I suspect there'll be a few objects that hang off BrowserContext/ResourceContext which are per profile that we don't want chrome to know about or even to create. For now, I did a one-off by manually keeping a map and deleting it when the BrowserContext goes away. If we hit others, we can generalize this into a template, perhaps like RenderViewHostObserverTracker.

BUG=98716
Review URL: https://chromiumcodereview.appspot.com/9384029

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@121706 4ff67af0-8c30-449e-8e8b-ad334ec8d88c