Fix chrome://omnibox/ to not throw exception when input is empty.
Precisely, the current version of the Javascript throws exceptions
if you check the show-incomplete-results checkbox and also if you play
with any checkboxes after that one is checked -- i.e., forcing a
call to refresh() -- if we don't have any output.  This fixes that.

BUG=
TEST=


Review URL: http://codereview.chromium.org/9385034

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@121717 4ff67af0-8c30-449e-8e8b-ad334ec8d88c