Allow SessionBackend::Init() to occur just-in-time in production builds.
This fixes a regression introduced in r113377 that caused the "old
session" to be switched at start-up before being used to populate the
recently closed menu.

This is the simplest possible fix, intended for merge to M18.  A more
robust fix that prevents future regressions is likely possible and may
be investigated as a follow-up.

BUG=110785
TEST=see bug repro steps

Review URL: https://chromiumcodereview.appspot.com/9361056

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@121878 4ff67af0-8c30-449e-8e8b-ad334ec8d88c