Disable accelerated compositing in component extensions
The policy for having accelerated compositing enabled for chrome WebUI should be consistent across the different types of URLs.  In particular, this works around an issue that was causing the open file manager dialog to be extremely slow on Alex devices (it's still quite slow, but nearly 2x faster with this change - see crbug.com/112628).

This change also removes a comment in related code that was stale (the code to which it refers was removed as part of TOUCH_UI cleanup).

TBR=ben@chromium.org (trivial comment fix)
BUG=105181
TEST=Run ChromeOS with the FPS counter enabled and verify that you don't get an FPS counter on Open file dialog windows.


Review URL: http://codereview.chromium.org/9385011

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@121888 4ff67af0-8c30-449e-8e8b-ad334ec8d88c