aura: Hide cursor during and after shutdown.
This updates the semantics of aura::RootWindow::ShowCursor()
so that after the cursor is hidden, it remains hidden during
subsequent calls to SetCursor().

It also makes us not show the cursor on mouse movement if
we're in the process of shutting down.

Finally, it makes us set an invisible cursor on the root
window when running on a Chrome OS device so that the cursor
will remain hidden after Chrome has exited during shutdown.

BUG=111884
TEST=manual: can't see cursor while shutting down now


Review URL: http://codereview.chromium.org/9388030

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@121960 4ff67af0-8c30-449e-8e8b-ad334ec8d88c