Linux: Overlay Chrome window icon with profile avatar or Incognito emblem/badge.
The window icon will consist of the usual Chrome product icon with the emblem
in the bottom-right corner.

BUG=108455
TEST=Manual. 1) Incognito window icons are always emblemed with the Incognito
icon, regardless of whether multi-profile is enabled or the number of
profiles. 2) With only one profile, regular window icons are not emblemed (but
Incognito ones are). 3) With more than one profile, all window icons are
emblemed: regular windows' icons with the current profile's avatar, and
Incognito window icons with the Incognito icon. 4) Icons from the current
screen's theme should take precedence over those packaged with Chrome. 5) If
no Chrome icons exist in the current screen's theme, the product icons
packaged with Chrome should be used. 6) If the user changes his profile
avatar, all of that profile's windows' icons should be updated to reflect the
change. 7) If the (system or Chrome) theme is changed, all windows' icons
should be updated to reflect the change. 8) App windows' icons consist of the
favicon and are emblemed if appropriate (see points 1-3, above). 9) Popup
windows' icons are emblemed if appropriate (see points 1-3, above).

Review URL: http://codereview.chromium.org/9363043

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@122161 4ff67af0-8c30-449e-8e8b-ad334ec8d88c