use a correct value for kShutdownDetectorThreadStackSize
Previously a hard coded value of 4096 was used as the stack size,
which is wrong because POSIX mandates that pthread_attr_setstacksize()
fails with EINVAL if the stacksize is less than PTHREAD_STACK_MIN.
On Linux it is smaller than PTHREAD_STACK_MIN, then it is set to RLIMIT_STACK,
which is huge compared to the expected 4096 bytes.
By using PTHREAD_STACK_MIN it is ensured that the smallest and still
correct stack size is used for the ShutdownDetectorThread.

BUG=
TEST=compile


Review URL: http://codereview.chromium.org/9405005

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@122182 4ff67af0-8c30-449e-8e8b-ad334ec8d88c