Add minimal CalculateIdleState implementation for Android.
It currently just reports that we're active on Android. Further
revisions could check device state to see if we're active but on memory
constrained devices such as Android we'll often get completely killed if
Chrome isn't active. This is a minimal step towards linking Chrome on
Android.

BUG=113487
TEST=


Review URL: http://codereview.chromium.org/9347052

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@122184 4ff67af0-8c30-449e-8e8b-ad334ec8d88c