Fix crash in sync promo dialog
Clicking on a link in the sync promo dialog was causing a crash.

The problem was that we were closing the dialog inside the click handler. This meant that any code in the handler after the close operation would crash if it referenced this.

Fix was to make the close operation asynchronous.

BUG=114391
TEST=


Review URL: http://codereview.chromium.org/9407025

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@122203 4ff67af0-8c30-449e-8e8b-ad334ec8d88c