aura: Clear screen for non-power-button exit.
This makes PowerButtonController clear the screen to black
when it hears that Chrome is about to exit.  We already
display an animation when the power button is held; this
handles other shutdown or signout paths like the wrench menu
or screen locker.

We listen for NOTIFICATION_APP_TERMINATING;
NOTIFICATION_APP_EXITING is sent too early (before
onbeforeunload handlers are called) and
NOTIFICATION_SESSION_END is sent too late.

BUG=108701,111884,113526
TEST=manual


Review URL: http://codereview.chromium.org/9402024

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@122224 4ff67af0-8c30-449e-8e8b-ad334ec8d88c