chromeos: use dbus::Properties in BluetoothAdapterClient
Add a Properties structure to BluetoothAdapterClient that encapsulates
the properties of an adapter object, manage the lifetime of this object
along with the ObjectProxy as a combined pair.

Switch to using a single PropertyChanged function for observers, which
can then get the Properties* for the passed in object path and look up
by name.

BUG=chromium-os:22087
TEST=built and tested on device

Change-Id: If4e60a84f8a9914b9155c232da7015dd7d41ee3c


Review URL: http://codereview.chromium.org/9414016

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@122454 4ff67af0-8c30-449e-8e8b-ad334ec8d88c