Add origin-based history removal to BrowsingDataRemover
This is the first of what will be many CLs to add origin-based deletion to
BrowsingDataRemover for all the variety of data types it supports. As a first
step, this CL hides origin-based functionality in the private
`BrowsingDataRemover::RemoveImpl()`. Once we're done with all the relevant data
types, we can add a `RemoveOrigin(int, GURL)` method next to `Remove(int)` to
expose a public API.

BUG=113965
TEST=unit_tests

Review URL: http://codereview.chromium.org/9379008

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@122497 4ff67af0-8c30-449e-8e8b-ad334ec8d88c