Implementation of a Matching strategy for URLs in the Declarative WebRequest API.
The Declarative WebRequest API needs to be able to match each outgoing
URLRequest against a large set of rules. This CL lays the foundation for this.
URLRequests can be matched using various criteria (e.g. "hostname ends in
foobar.com", or "URL contains 'secret'", or both). In order to match many of
these patterns very quickly we want to rely on the Aho-Corasick algorithm.

This CL is one step before the A.-C. algorithm: It implements the reduction of
URL Pattern matches to that of Substring matches. The SubstringSetMatcher is
implemented trivially and will be replaced with a A.-C. implementation in a
second CL.

BUG=112155
TEST=no


Review URL: http://codereview.chromium.org/9390018

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@122503 4ff67af0-8c30-449e-8e8b-ad334ec8d88c