Adds dump accessibility tree support in Windows.
This largely adds human readable strings for IAccessible2 roles and states.
This is in part to make it possible to hand-edit the expected files we generate.
Sidenote: switch to string16 usage from std::string as much as possible (despite the incompatibility with API's provided by FilePath).

Review URL: https://chromiumcodereview.appspot.com/9295031

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@122617 4ff67af0-8c30-449e-8e8b-ad334ec8d88c