chromeos: use BluetoothDeviceClient::Properties for DeviceFound
Use the newly defined BluetoothDeviceClient::Properties structure for the
BluetoothAdapterClient::Observer::DeviceFound call, rather than a
dictionary of arbitrary values obtained from the remote process.

This is not only more type-safe but means we can be utterly untrusting
of the remote data, and only take the pieces we're looking for, rather
than passing everything to the JS UI handler.

BUG=chromium-os:22087
TEST=built and tested on device

Change-Id: If1f797ef37856e98f9443df04243ae67980cc7f9


Review URL: http://codereview.chromium.org/9418028

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@122703 4ff67af0-8c30-449e-8e8b-ad334ec8d88c