Handle blur events from user pod elements.
Avoid having the appearance of focus when we do not have focus. Clicking on the menu removes focus from the user pod but visually does not do so. This should make it less painful in the cases where the lock screen password element loses focus on resume as well.

BUG=114810
TEST=Click on a status area menu, and when the blur event is sent to the password element (or guest enter button) the pod loses focus.

Review URL: http://codereview.chromium.org/9423011

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@122765 4ff67af0-8c30-449e-8e8b-ad334ec8d88c