Change tracking of source of sync cycles
Since r121290 (issue 111229), we no longer depend on
SYNC_CYCLE_CONTINUATION to decide how to handle failing sync cycles.  As
far as I know, this was the only piece of client code that made
decisions based on the sync cycle source.  The sole remaining purpose of
the updates_source variable is debugging.  This presents an opporunity
to increase the usefulness of this variable without affecting any client
behaviour.

Whereas we would normally set the source to SYNC_CYCLE_CONTINUATION in
the DOWNLOAD_UPDATES step, this commit delays resetting the variable
until later.  The practical effect of this is that the source reported
by snapshots (and exposed in about:sync) can now have values other than
SYNC_CYCLE_CONTINUATION.  This means about:sync now properly reports if
the most recent sync cycle was a due to a local change, a notification,
etc.

The SYNC_CYCLE_CONTINUATION variable is still used to identify sync
cycles that are run because the original sync cycle left some tasks
unfinished.  It is up to SyncSession::HasMoreToSync() to decide when
these cycles are necessary.

We should note that the updates_source is reported to the server for
debug purposes, and that this commit will affect the values we report.
Whereas we used to report jobs tried from backoff as having source
SYNC_CYCLE_CONTINUATION, this change will allow their original source to
remain intact.  A series of attempts to contact the server while in
backoff should all have the same source once this patch is applied.

BUG=98346
TEST=


Review URL: http://codereview.chromium.org/9375032

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@122937 4ff67af0-8c30-449e-8e8b-ad334ec8d88c