ash: Allow dragging using touch/gesture events.
This is the first step towards making it possible to drag with touch/gesture events. There will be subsequent CLs to make sure the correct touch-id is used for dragging, and update the drag-image position to be slightly above the touch-point so that it doesn't get obscured by the touch-finger.

BUG=114752
TEST=mouse-event dragging should be covered by existing tests. Tests for touch-events will be added in subsequent CLs.

Review URL: https://chromiumcodereview.appspot.com/9418043

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@122982 4ff67af0-8c30-449e-8e8b-ad334ec8d88c