Fix chrome crash on clicking Scratchpad's "Send to tab"
The crash happens when scratchpad app tries transision from panel window to tabbedd window.
When the "Send to tab" button is clicked, the app closes the panel window and right after that it tries to open itself in a new tab (by calling tabs.create() in callback from tabs.remove(panel_tab).

The problem is that when we try to create new tab, panel browser window is still in browser list (eventhough it is being closed), and it is the browser returned by GetCurrentBrowser.
So, we end up trying to create a new tab in the panel window that is just about to go away (browser navigator does not exchange the params.browser with a tabbed browser because the panel browser has empty tabstrip_model, so creating new tab is allowed)

BUG=chromium-os:25804
TEST=confirmed that crash is not happening locally (on M18 chrome with this patch applied)

Review URL: https://chromiumcodereview.appspot.com/9403032

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@123056 4ff67af0-8c30-449e-8e8b-ad334ec8d88c