Change how you get to the MessageLoop::Dispatcher.
Previously, you would ask the RootWindow (a singleton) for its dispatcher, which was conveniently the RootWindowHost. Now that we are moving towards many RootWindows, you must ask the new singleton, aura::Env, instead.

This CL also implements a dispatcher for windows. It does not implement the Linux one since that's much larger and I'll do that in a followon CL. I just wanted to change the API. Note there is some transitional ugliness.

http://crbug.com/112131
TEST=existing
Review URL: https://chromiumcodereview.appspot.com/9433072

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@123268 4ff67af0-8c30-449e-8e8b-ad334ec8d88c