[uber] set line-height: 1 on header elements
This makes the height of <h*> independent of the choice of font family (this is important because different platforms might have different fonts, but we don't want that to affect spacing).

I've changed spacing around most header elements to compensate for new visual differences.

BUG=110760
TEST=manual


Review URL: http://codereview.chromium.org/9452003

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@123429 4ff67af0-8c30-449e-8e8b-ad334ec8d88c