ash: A first step for the uber tray.
As a first step, there are items to open settings, help pages, and an item to
allow [un]muting volume.

The resources have not been uploaded yet. That will be done in a separate CL to
allow this one to go through trybots etc.

This new uber-tray can be turned on at runtime using --ash-uber-tray flag.

BUG=110130
TEST=none yet. will be added when a little more fleshed out.

Review URL: https://chromiumcodereview.appspot.com/9463039

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@123799 4ff67af0-8c30-449e-8e8b-ad334ec8d88c