Change handling of clientInfo.numericDate during about:net-internals imports.
This makes the field required, so other code can assume its existence.

For dump files which do not contain the property (either because they predated M17, or because they were generated through the command-line logging), we synthesize a value during import. The synthesized value is simply the timestamp of the latest event.

Review URL: http://codereview.chromium.org/9515012

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@124136 4ff67af0-8c30-449e-8e8b-ad334ec8d88c