[Sync] Dont re-encrypt data when new nigori arrives.
The logic was previously meant to handle newly encrypted types. But, it would
also attempt to re-encrypt data when any new nigori arrived. We now skip
data that is already encrypted (even if it's not with the most recent
encryption keys).

BUG=115309
TEST=sync_unit_tests. Receiving a nigori node update that leaves the
cryptographer ready for an account with enrypt everything should no longer
crash.


Review URL: http://codereview.chromium.org/9550001

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@124354 4ff67af0-8c30-449e-8e8b-ad334ec8d88c