GTK: Draw the correct omnibox state with new tab-to-traverse behavior.
This does everything but animate the trailing ">" icon from the right
side of the omnibox to the left when entering tab-to-search mode. The
TTS arrow indicator is now properly GTK themed, as well.

(Doing the animation is lower priority. It would require more
extensive refactoring that I don't have time for as I have code yellow work.)

BUG=116292
TEST=all new tab-to-search behavior matches views sans animation.


Review URL: http://codereview.chromium.org/9562029

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@124552 4ff67af0-8c30-449e-8e8b-ad334ec8d88c