Refactor this test to reduce code duplication.
This was done as a precursor to implementing a TDD test for the bug 111376. I'm noting the bug number here in case the follow-up revision is merged to M17 or M18 (in which case this refactoring might be required to make the tests compile).


BUG=111376
TEST=unit_tests --gtest_filter=SQLitePersistentCookieStoreTest.*


Review URL: http://codereview.chromium.org/9569028

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@124650 4ff67af0-8c30-449e-8e8b-ad334ec8d88c