Add an android specific ntp resource cache.
This is a minimal starting point for merging our upstream android
ntp resource changes into a somewhat clean slate.  Based on how little of the
shared resource cache code we used, it seems cleaner to start our own and
not muddle that file with a ton of ifdefs.

This is all in the hopes of getting sync_unit_tests linking upstream.

BUG=113487
TEST=


Review URL: http://codereview.chromium.org/9558014

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@124767 4ff67af0-8c30-449e-8e8b-ad334ec8d88c