Aura: Fix gray line at top of toolbar with translucent frames
The toolbar background painting code was copying pixels out of the toolbar background image with the wrong vertical offset.  I also moved the paint code into BrowserNonClientFrameViewAura so it can get the proper tabstrip vertical offset, and also because that's a more appropriate place for the logic.

BUG=113075
TEST=visual, ensure no "step" in the gradient at the bottom of the active tab or the top of the toolbar, for both maximized and restored windows
TBR=ben@chromium.org


Review URL: http://codereview.chromium.org/9583032

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@124836 4ff67af0-8c30-449e-8e8b-ad334ec8d88c