gdata:: Add GDataFileSystem::GetFile().
This function is used to get files from the file system.
For now, it always download files from the cloud.
We should have the caching layer in-pace here.

Along the way, a mssing 'return' to Remove().
Change GetDocumentUrlFromPath() to GetGDataFileInfoFromPath,
and wrote a test for it.

BUG=chromium-os:27174
TEST=unit_tests


Review URL: http://codereview.chromium.org/9583031

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@124857 4ff67af0-8c30-449e-8e8b-ad334ec8d88c