Changes related to new conflict counters
While it might have made more sense to commit these separately, it's
hard to update the snapshot format without leaving some part of the
system in an inconsistent state.

At the heart of this change is a change in the snapshot to make it contain
counts of various types of conflicts, rather than just counting the simple
conflicts and total number of conflicts.  This enables us to make the changes
that follow.

This change adds more detailed conflict information to about:sync.  The main
page now lists the counts of encryption, hierarchy, simple and server conflicts
separately.  It removes the total number of conflicts counter, but that
number should be equal to encryption + hierarchy + simple conflicts.

Similarly, it updates the client_debug_info server-side debug mechanism with
the new counters.  The old counters are not only old, but were also being set
wrong.  I've added comments to the protobuf that explain the situation.  We
should move away from those old counters as soon as possible.

Finally, it removes the conflicting_commits counter from ErrorCounters.
There's a new server conflicting commits counter (intended to be symmetric with
the other new counters) that can track this.  Exposing that counter has made
conflicting_commits redundant.

BUG=107816,112659
TEST=


Review URL: http://codereview.chromium.org/9428016

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@125287 4ff67af0-8c30-449e-8e8b-ad334ec8d88c