gdata: First cut at adding support for uploading from gdata.
Known issue 1: the file name shown in the web page, and used for
uploading is human-unreadable and wrong (as of now, it looks like
".org.chromium.Chromium.CtC9KF").
crosbug.com/27222

Known issue 2: the downloading progress is not shown in the UI.
crosbug.com/27504

Known issue 3: downloading cannot be canceled.
crosbug.com/27505

Known issue 4: the downloaded file is not deleted. The issue is
gone once the cache layer in GDataFileSystem is in place.
crosbug.com/26910

BUG=chromium-os:27244
TEST=manually

Review URL: https://chromiumcodereview.appspot.com/9640001

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@125645 4ff67af0-8c30-449e-8e8b-ad334ec8d88c