ash: Fix unlocking screen after locking when using the uber-tray.
The tray-delegate has a LockScreen function that is used to lock the screen. It
turns out, PowerManagerClient::Observer has an observer callback, also called
LockScreen, which gets called when the screen is locked. As a result, when the
user locks the screen, the observer's LockScreen gets called again, and it falls into a cycle.

In short, you dawg, I heard you like to lock your screen, so I locked your screen
while I tell you that I locked your screen.

Renamed to RequestLockScreen to avoid the collision.

BUG=110130
TEST=none

Review URL: https://chromiumcodereview.appspot.com/9649008

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@125724 4ff67af0-8c30-449e-8e8b-ad334ec8d88c