bluetooth: fix issues with device removal
The DeviceRemoved signal comes with the object path, rather than the
address, but GetProperties can't be called to get the address because
the device has gone. Iterate the devices comparing object paths to
find the one we need; this is short and rare enough that a map would
be more wasteful.

Also the BrowserOptions.removeBluetoothDevice() function added in
I9624018 takes a string address rather than a device dictionary, so call
that correctly.

BUG=chromium-os:21320
TEST=verified on device after switching off mouse during discovery

Change-Id: Ie073282f92d3215a2e4a9a68b06d770ee65307cc


Review URL: http://codereview.chromium.org/9648010

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@125803 4ff67af0-8c30-449e-8e8b-ad334ec8d88c