dbus: remove service name from ExportedObject
Well-known names in D-Bus are merely aliases to unique connection ids
maintained by the bus, they have no purpose in qualifying object paths
or interfaces and it's perfectly legimiate for a client to make
requests to the unique connection id (e.g. in response to a signal,
which does not reference the well-known name of the origin connection).

Remove the service_name member from dbus::ExportedObject, from its
constructor and from dbus::Bus::GetExportedObject and require code to
call dbus::Bus::RequestOwnership if a well-known name is desired. This
requires making that function callable from the origin thread with
a callback for the return value.

BUG=chromium-os:27101
TEST=dbus_unittests

Change-Id: Ib91de8b68ad9c3b432e224a2c715f0c2ca1af463


Review URL: http://codereview.chromium.org/9668018

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@125970 4ff67af0-8c30-449e-8e8b-ad334ec8d88c