Provide a policy option to enable online revocation checks.
For Chrome 19, when online OCSP/CRL are disabled, this policy option provides
enterprises the ability to restore the old behaviour.

We have also had a request for an option to enable hard-fail checks which would
effectively make this preference a tri-state: off, soft, hard. However, I've
elected to make this a bool because I think that two bools will be easier to
understand than a string or magic number.

BUG=none
TEST=Set EnableOnlineRevocationChecks to true and observe OCSP checks in the net-internals log when navigating to https://www.bankofamerica.com.


Review URL: http://codereview.chromium.org/9578020

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@126167 4ff67af0-8c30-449e-8e8b-ad334ec8d88c