Implement ephemeral users
This CL completes the implementation of http://crosbug.com/20004.
When the ephemeral users policy is enabled, users are not added to
UserManager's persistent list and are not asked to pick a user image.
UserManager additionally checks for stale entries created before the
policy was enabled and removes these, along with downloaded user
images.

See http://goo.gl/wTPEb for a design doc.

The CL includes unit tests for the new feature.

BUG=chromium-os:20004
TEST=unittest UserManagerTest*
TEST=manual feature test in numerous scenarios on a Chromebook

Review URL: http://codereview.chromium.org/9405035

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@126198 4ff67af0-8c30-449e-8e8b-ad334ec8d88c